Fixing some issues
* Fixing bad .type property in get where the selector is a function
* Ensuring all specs pass
* Adding stub for hooks spec for the time being...
