+++ lib/index.ts
@@ -1,5 +1,7 @@
 /// <reference path="./index.d.ts" />
 
+import getDisplayName from './getDisplayName';
+
 // having weak reference to styles prevents garbage collection
 // and "losing" styles when the next test starts
 const stylesCache = new Map()
@@ -16,57 +18,6 @@ const setXMLHttpRequest = w => {
 const setAlert = w => {
   window.alert = w.alert
   return w
-}
-
-const cachedDisplayNames: WeakMap<JSX, string> = new WeakMap();
-
-/**
- * Gets the display name of the component when possible.
- * @param type {JSX} The type object returned from creating the react element.
- * @param fallbackName {string} The alias, or fallback name to use when the name cannot be derived.
- * @link https://github.com/facebook/react-devtools/blob/master/backend/getDisplayName.js
- */
-function getDisplayName(type: JSX, fallbackName: string = 'Unknown'): string {
-  const nameFromCache = cachedDisplayNames.get(type)
-
-  if (nameFromCache != null) {
-    return nameFromCache
-  }
-
-  let displayName: string
-
-  // The displayName property is not guaranteed to be a string.
-  // It's only safe to use for our purposes if it's a string.
-  // github.com/facebook/react-devtools/issues/803
-  if (typeof type.displayName === 'string') {
-    displayName = type.displayName
-  }
-
-  if (!displayName) {
-    displayName = type.name || fallbackName
-  }
-
-  // Facebook-specific hack to turn "Image [from Image.react]" into just "Image".
-  // We need displayName with module name for error reports but it clutters the DevTools.
-  const match = displayName.match(/^(.*) \[from (.*)\]$/)
-
-  if (match) {
-    const componentName = match[1]
-    const moduleName = match[2]
-
-    if (componentName && moduleName) {
-      if (
-        moduleName === componentName ||
-        moduleName.startsWith(componentName + '.')
-      ) {
-        displayName = componentName
-      }
-    }
-  }
-
-  cachedDisplayNames.set(type, displayName)
-
-  return displayName
 }
 
 /** Initialize an empty document w/ ReactDOM and DOM events.
@@ -217,7 +168,7 @@ Cypress.Commands.overwrite('get', (originalFn, selector, options) => {
     case 'function':
       // If attempting to use the component name without JSX (testing in .js/.ts files)
       // const displayname = selector.prototype.constructor.name
-      const displayname = getDisplayName(selector.type);
+      const displayname = getDisplayName(selector);
       return originalFn(`@${displayname}`, options)
     default:
       return originalFn(selector, options)
